+++ integrated_channels/cornerstone/__init__.py
@@ -3,6 +3,3 @@ The Cornerstone Integrated Channel package.
 """
 
 __version__ = "0.1.0"
-
-default_app_config = ("integrated_channels.cornerstone.apps."
-                      "CornerstoneConfig")