+++ integrated_channels/degreed/admin/__init__.py
@@ -76,6 +76,9 @@ class DegreedEnterpriseCustomerConfigurationAdmin(DjangoObjectActions, admin.Mod
         """
         return obj.enterprise_customer.name
 
+    @admin.action(
+        description="Force content metadata transmission for this Enterprise Customer"
+    )
     def force_content_metadata_transmission(self, request, obj):
         """
         Updates the modified time of the customer record to retransmit courses metadata
@@ -101,9 +104,6 @@ class DegreedEnterpriseCustomerConfigurationAdmin(DjangoObjectActions, admin.Mod
         )
 
     force_content_metadata_transmission.label = "Force content metadata transmission"
-    force_content_metadata_transmission.short_description = (
-        "Force content metadata transmission for this Enterprise Customer"
-    )
 
 
 @admin.register(DegreedLearnerDataTransmissionAudit)