+++ tests/test_integrated_channels/test_canvas/test_utils.py
@@ -10,14 +10,12 @@ from unittest import mock
 import pytest
 from requests.models import Response
 
-from django.utils import timezone
-
 from integrated_channels.canvas.utils import CanvasUtil
 from integrated_channels.exceptions import ClientError
 from integrated_channels.utils import refresh_session_if_expired
 from test_utils import factories
 
-NOW = datetime.datetime(2017, 1, 2, 3, 4, 5, tzinfo=timezone.utc)
+NOW = datetime.datetime(2017, 1, 2, 3, 4, 5, tzinfo=datetime.timezone.utc)
 NOW_TIMESTAMP_FORMATTED = NOW.strftime('%F')