+++ tests/test_integrated_channels/test_degreed/test_client.py
@@ -12,13 +12,11 @@ import requests
 import responses
 from freezegun import freeze_time
 
-from django.utils import timezone
-
 from integrated_channels.degreed.client import DegreedAPIClient
 from integrated_channels.exceptions import ClientError
 from test_utils import factories
 
-NOW = datetime.datetime(2017, 1, 2, 3, 4, 5, tzinfo=timezone.utc)
+NOW = datetime.datetime(2017, 1, 2, 3, 4, 5, tzinfo=datetime.timezone.utc)
 NOW_TIMESTAMP_FORMATTED = NOW.strftime('%F')